task: split out functions
